Rename Private favourites package for KSP compatibility
